AnimUtils : clear listeners when animation completes or is terminated
A few of the functions in AnimUtils set listeners on the ViewPropertyAnimator
of the view they are animating. These listeners are never cleared after the
animation is over or cancelled. They linger around and can be inadvertently
triggered on another un-related animation. This change ensures that if a
listener is set for a particular animation, it is cleared once the
animation completes.

Change-Id: I0f1948d878f160ac3102e5cb46347295978a57fe